requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
paramiko==3.5.1
jinja2==3.1.6
pyyaml==6.0.2
httpx==0.28.1
itsdangerous==2.2.0
python-multipart==0.0.20
